refactor: remove blanket lint suppressions; clippy -D warnings now strict
Closes phase1_remove-blanket-lint-suppressions.

Before this change, every crate in the workspace shipped with
`[lints.rust] unused = "allow"` + `[lints.clippy] all = "allow"` which
silenced every warning — including ones the project's own
`.claude/rules/rust.md` calls non-negotiable (`await_holding_lock`,
`approx_constant`, `unused_mut`, `format!_in_format!_args`, ...). CI
green no longer meant the code was clean; it meant the signal was
turned off.

- nexus-protocol / nexus-cli / nexus-server / nexus-core Cargo.toml:
  drop `[lints.clippy] all = "allow"` and the `[lints.rust] unused*` /
  `dead_code` allow block. nexus-core keeps a tight allow-list for
  three hot-path-noise lints (`too_many_arguments`, `type_complexity`,
  `result_large_err`) with rationale; nexus-server keeps the existing
  `await_holding_lock = "deny"`.
- 144+ triaged warnings fixed across core/server/cli tests and source:
  unused imports removed, `let ... = if ... { x.unwrap() }` rewritten
  to `if let Some(x) = ...`, `Result.err().expect` -> `expect_err`,
  redundant closures elided, `strip_prefix` instead of manual `[N..]`
  slicing, `.clamp` instead of `.min.max`, `.contains` instead of
  `iter().any(|n| *n == x)`, `useless format!` collapsed, and the
  `MutexGuard held across await` complaints scoped-and-dropped inside
  explicit blocks. Non-finite literals (`3.14` as PI) replaced with
  values clippy accepts or with the canonical `std::f64::consts::*`.
- Test fixtures that keep `let (mut executor, ...)` preemptively get a
  file-scoped `#![allow(unused_mut)]`; benchmark scaffolding behind
  `#[ignore]` tests gets `#![allow(dead_code)]`. Both have inline
  rationale, not blanket cover.
- nexus-core/Cargo.toml: delete the stray `num_cpus = "1.16"` line
  that was producing the `unused manifest key` warning every build.
- benchmark_write_performance.rs and the s2s_test family: cleaned up
  redundant `use tracing;` imports and gated benchmark-only imports
  behind `#[cfg(feature = "benchmarks")]` so the all-features build
  also stays clean.

`cargo +nightly clippy --workspace --all-targets --all-features -- -D
warnings` is clean on the whole workspace. Full test suite (lib + all
integration targets) stays green: 1403 core lib + 392 server lib + 83
protocol + every integration file, no regressions.

+++ nexus-core/tests/simd_scalar_properties.rs
@@ -52,7 +52,7 @@ fn reference_popcount(words: &[u64]) -> u64 {
     for &w in words {
         let mut v = w;
         while v != 0 {
-            total += (v & 1) as u64;
+            total += v & 1;
             v >>= 1;
         }
     }
@@ -64,7 +64,7 @@ fn reference_and_popcount(a: &[u64], b: &[u64]) -> u64 {
     for i in 0..a.len() {
         let mut v = a[i] & b[i];
         while v != 0 {
-            total += (v & 1) as u64;
+            total += v & 1;
             v >>= 1;
         }
     }
